requirements.txt
Django==2.2.7
django-bootstrap3==11.1.0
django-jquery==3.1.0
django-registration==3.0.1
djangorestframework==3.8.2
psycopg2==2.8.4
psycopg2-binary==2.8.4
pytz==2019.3
requests==2.22.0
django-rest-auth==0.9.5
django-cors-headers==3.1.1
django-rest-swagger==2.2.0
gunicorn==20.0.0